aloisioimoveis/locations/serializers.py
from rest_framework import serializers

from aloisioimoveis.locations.models import City, Neighborhood


class CitySerializer(serializers.ModelSerializer):
    class Meta:
        model = City
        fields = ('id', 'name')


class NeighborhoodSerializer(serializers.ModelSerializer):
    class Meta:
        model = Neighborhood
        fields = ('id', 'name')
