requirements.txt
cloudinary==1.8.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.2
django-debug-toolbar==1.8
django-extensions==1.7.9
django-rainbowtests==0.6.0
djangorestframework==3.6.3
django-test-without-migrations==0.6
gunicorn==19.7.1
model_mommy==1.3.2
psycopg2==2.7.1
python-decouple==3.0
pytz==2017.2
static3==0.7.0
